Revert "like/dislike bookmark"
