+++ src/app/store.js
@@ -1,16 +1,12 @@
 import { configureStore } from "@reduxjs/toolkit";
-import {
-  authReducer,
-  bookmarkReducer,
-  postReducer,
-  userReducer,
-} from "../features";
+import authReducer from "../features/auth/authSlice";
+import userReducer from "../features/user/userSlice";
+import postReducer from "../features/post/postSlice";
 
 export default configureStore({
   reducer: {
     auth: authReducer,
     user: userReducer,
     post: postReducer,
-    bookmark: bookmarkReducer,
   },
 });
