--- src/features/index.js
@@ -1,11 +1,0 @@
-import authReducer from "./auth/authSlice";
-import userReducer from "./user/userSlice";
-import postReducer from "./post/postSlice";
-import bookmarkReducer from "./bookmark/bookmarkSlice";
-
-export {
-  authReducer,
-  userReducer,
-  postReducer,
-  bookmarkReducer
-}
