--- src/helper/index.js
@@ -1,29 +1,0 @@
-import { loginHandler, signUpHandler } from "./auth";
-import { bookmarkPost } from "./bookmark/bookmarkPost";
-import { unsavePost } from "./bookmark/unsavePost";
-import {
-  editPost,
-  likePost,
-  sendPost,
-  deletePost,
-  dislikePost,
-  getAllPosts,
-} from "./post";
-import { getAllUsers, followUser, unfollowUser, editUser } from "./user";
-
-export {
-  editPost,
-  likePost,
-  sendPost,
-  editUser,
-  unsavePost,
-  deletePost,
-  followUser,
-  dislikePost,
-  getAllPosts,
-  getAllUsers,
-  unfollowUser,
-  bookmarkPost,
-  loginHandler,
-  signUpHandler,
-};
