+++ src/helper/post/index.js
@@ -1,14 +1,10 @@
 import { deletePost } from "./deletePost";
-import { dislikePost } from "./dislikePost";
 import { editPost } from "./editPost";
 import { getAllPosts } from "./getAllPosts";
-import { likePost } from "./likePost";
 import { sendPost } from "./sendPost";
 export { 
   editPost,
-  likePost,
   sendPost,
   deletePost,
-  dislikePost,
   getAllPosts,
  };
